+++ source/base/PlatformUtilities.c
@@ -93,7 +93,7 @@ CharString getPlatformName(void) {
   Gestalt(gestaltSystemVersionMajor, &major);
   Gestalt(gestaltSystemVersionMinor, &minor);
   Gestalt(gestaltSystemVersionBugFix, &bugfix);
-  snprintf(result->data, result->length, "Mac OS X %ld.%ld.%ld", major, minor, bugfix);
+  snprintf(result->data, result->length, "Mac OS X %d.%d.%d", (int)major, (int)minor, (int)bugfix);
 #elif LINUX
   CharString line = newCharString();
   char *lineDelimiter = NULL;
